webapi/webapi-mediaserver-tizen-tests/mediaserver/MediaContainerIdEvent_id_value.html

<!DOCTYPE html>
<!--
Copyright (c) 2014 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Wang, Hongjuan <hongjuanx.wang@intel.com>

-->

<meta charset="utf-8">
<title>DLNA Test: MediaContainerIdEvent - id value</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://01org.github.io/cloud-dleyna/mediaserver.html#mediacontaineridevent-interface">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<div id='serverContainer'></div>
<script>
  var t = async_test("Check if the MediaContainerIdEvent.id MUST return the id of the MediaContainer that triggered the containerchanged event when getting");
  var t1 = async_test("the onserverfound event can be invoked");
  var msga = navigator.mediaServer;
  msga.scanNetwork();

  function onFolderCreated(result) { }

  function onCreateFailed(error) {
    t.step(function () {
      assert_unreached("Error message: { " + error.message + " }");
    });
    t.done();
  }

  msga.onserverfound = t1.step_func( function (event) {
    var mse = event.server;
    var mc = mse.root;
    var container = document.getElementById('serverContainer');
    var createContainerField = document.createElement('input');
    var p = document.createElement('p')
    createContainerField.id = 'createContainerField';
    createContainerField.serverId = event.server.id;
    container.serverId = createContainerField.serverId;
    container.appendChild(p);
    p.appendChild(createContainerField);
    createContainerField.value = "createContainer";
    var title = createContainerField.value;
    mc.createFolder(title).then(onFolderCreated, onCreateFailed);
    mse.oncontainerchanged = t.step_func( function (eve) {
      assert_equals(eve.id, mc.id, "the MediaContainerIdEvent.id value");
      t.done();
    });
    t1.done();
  });
</script>
